Creating scipy sparse matrix does not copy data
Turns out that scipy does not copy the input arrays when creating a sparse matrix explicitely (e.g.
from compressed values and indices). This means it wraps around the C arrays quite easily.

Changed docstring and test to reflect that.
However, current API is not safe, since the C memory is not handled by the python API. This can
result in numpy arrays with dangling data pointers. We should create a cython to explicitely manage
_SparseMatRow's memory.