Initializing RNG with seed in example_30dor.c
The SOPT random number generator needs to be initialized by calling it with -seed as argument.
This was not done in example_30dor.c. Adding the initialization bit  makes it possible to compare
example_30dor.c quantities to those generated by the python bindings, since it provides an initial
state that the RNG can be reset to.