Adds SOPT RNG bindings for tests only
We want the same random numbers in both C and python when testing. This
requires python bindings for the SOPT random number generator. However, it is
better to use the many RNG tools out there, rather than the in-house one. So
this this binding is only available in testing.

Note: c144c8fddef5ec8d4d4cd53e50b587dccf479c45 makes sure the state of the
      random number generator can be set and reset