Make voided data accessible from cython extensions
The C functions need extra data passed via a `void *` array.
The MeasurementOperator and SparsityOperator declare functions that take care
of this, so that the voided can be used from other extension types (namely the
solvers).

The MeasurementOperator especially requires an array of 5 void pointer to 5
different data objects. A specific extension type makes sure that this array
contains live objects that stay live.