PyArray_DATA rather than generic buffer interface
for the sake of older numpys without the newer python buffer interface.
The main issue is accessing pointers to the underlying data for
both complex and real arrays.
This is what the buffer interface allows us to do.
On the other hand, the normal cython numpy interface expects an explicit
type. It is simpler to roll out are own typeless for those arrays that
could be either complex or real.
In a few years / CASA upgrades, it may be forthwhile to replace all calls to
this interface with the normal cython buffer interface. One should look at the
changeset when this particular file was introduced.

In the same vain, numpy arrays are created explicitly with order="C".