Installation with CASA 4.2.2 on mac
Partially addresses #45. The purify installer calls install_name_tool so
libpython2.7.dylib is referred to with an absolute path. It is not clear why it
isn't absolute in the first place since it appears with an absolute path on the
commandline during linking. Might be because it is part of a python framework
belonguing to the CASA app. Mysterious are the ways of the Mac.